# PERF: Optimize Base64Convert.cs