--- a/Base64Extensions.Benchmarks/EncodeBenchmark.cs
+++ b/Base64Extensions.Benchmarks/EncodeBenchmark.cs
@@ -1,6 +1,5 @@
 ﻿using BenchmarkDotNet.Attributes;
 using System;
-using System.Security.Cryptography;
 
 namespace Base64Extensions.Benchmarks
 {
@@ -15,7 +14,7 @@ namespace Base64Extensions.Benchmarks
         {
             _data = new byte[DataLength];
 
-            var rng = new RNGCryptoServiceProvider();
+            var rng = new DeterministicRandomGenerator();
             rng.GetBytes(_data);
         }