--- a/Base64Extensions.Benchmarks/Program.cs
+++ b/Base64Extensions.Benchmarks/Program.cs
@@ -1,7 +1,18 @@
 ﻿using BenchmarkDotNet.Running;
+using System;
+using System.Security.Cryptography;
 
 namespace Base64Extensions.Benchmarks
 {
+    class DeterministicRandomGenerator : RandomNumberGenerator
+    {
+        Random r = new Random(0);
+        public override void GetBytes(byte[] data)
+        {
+            r.NextBytes(data);
+        }
+    }
+
     public class Program
     {
         public static void Main(string[] args)