Hi @reecerussell, I've tried to make some performance optimizations to this repo. Following are the benchmark results before and after the changes:

**Benchmarks before changes:**

|                   Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| EncodeUsingBase64Convert | 147.8 ns | 3.03 ns | 7.48 ns | 0.3519 |     - |     - |     736 B |
| EncodeUsingSystemConvert | 429.8 ns | 5.78 ns | 5.41 ns | 0.3405 |     - |     - |     712 B |

|                   Method |      Mean |     Error |    StdDev |    Median |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------- |----------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
| DecodeUsingBase64Convert | 15.286 us | 0.3408 us | 0.9889 us | 14.890 us | 2.5330 |     - |     - |    5328 B |
| DecodeUsingSystemConvert |  1.011 us | 0.0118 us | 0.0110 us |  1.011 us | 0.1335 |     - |     - |     280 B |

**Benchmarks after changes:**
|                   Method |     Mean |    Error |   StdDev |    Median |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------- |---------:|---------:|---------:|----------:|-------:|------:|------:|----------:|
| EncodeUsingBase64Convert | 101.0 ns |  3.69 ns | 10.42 ns |  98.35 ns | 0.1760 |     - |     - |     368 B |
| EncodeUsingSystemConvert | 452.3 ns | 14.43 ns | 41.41 ns | 438.71 ns | 0.3405 |     - |     - |     712 B |

|                   Method |      Mean |     Error |    StdDev |    Median |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------- |----------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
| DecodeUsingBase64Convert | 15.041 us | 0.2836 us | 0.6517 us | 14.820 us | 2.5330 |     - |     - |    5328 B |
| DecodeUsingSystemConvert |  1.105 us | 0.0219 us | 0.0553 us |  1.099 us | 0.1335 |     - |     - |     280 B |

I've also made sure that all the unit tests pass following these changes as well. Could you please confirm whether or not you think my changes are valid?

Thank you!